pyproject.toml
[project]
name = "autotrader"
version = "0.1.0"
description = "Autonomous trading strategy research (autoresearch for Hyperliquid)"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "scipy>=1.11.0",
    "requests>=2.31.0",
    "pyarrow>=14.0.0",
    "websockets>=12.0",
    "httpx>=0.25.0",
    "eth-account>=0.10.0",
    "python-dotenv>=1.0.0",
    "aiosqlite>=0.19.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.0",
    "hyperliquid-python-sdk>=0.22.0",
    "loky>=3.5.6",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.ruff.lint]
ignore = ["E402"]  # sys.path manipulation before imports is intentional

[tool.ruff.lint.per-file-ignores]
"**/__init__.py" = ["F401"]  # re-exports
"**/adapter/ensemble.py" = ["F401"]  # re-export alias
